+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- remove the '--save' flag from `bit import`, it always saves the imported components to the bit.json
 - do not cache components in the objects directory upon `bit import`, unless it is a bit bare scope. 
 - add the commit-log details to the Component object, so then it'll be available for `bit show --json` and `bit export`.